Break app.tsx to different files and rearrange 'pages/' structure
This implements the extraction of Header and removal of Footer from
app.tsx. Also rearranges the folder structure for 'pages', making
necessary changes to TS and SASS imports.
