+++ client/src/app/header.scss
@@ -1,0 +1,14 @@
+.nav {
+    @extend .padding8;
+    background-color: $mainColor;
+    .title a,
+    .is-hidden-mobile .nav-item {
+        color: white;
+    }
+    .nav-toggle span {
+        background-color: white;
+    }
+    .nav-toggle:hover {
+        background-color: transparent;
+    }
+}
