+++ client/src/page/createProfile/createProfile.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import {connect} from 'react-redux'
-import {State} from '../redux/reducers'
-import {pick} from '../utils/utils'
-import * as selectors from '../redux/selectors'
+import {State} from '../../redux/reducers'
+import {pick} from '../../utils/utils'
+import * as selectors from '../../redux/selectors'
 
 interface Props {
   createProfileUrlInputValue: string
