+++ client/src/page/homepage/homepage.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react'
 import {connect} from 'react-redux'
-import {actions} from '../redux/actions'
-import {State} from '../redux/reducers'
-import {pick} from '../utils/utils'
-import {Form} from '../component/form'
-import * as selectors from '../redux/selectors'
-import {FadeInUp} from '../component/animations'
-import {IAlgoliaSearchResults} from '../vendor/algolia'
+import {actions} from '../../redux/actions'
+import {State} from '../../redux/reducers'
+import {pick} from '../../utils/utils'
+import {Form} from '../../component/form'
+import * as selectors from '../../redux/selectors'
+import {FadeInUp} from '../../component/animations'
+import {IAlgoliaSearchResults} from '../../vendor/algolia'
 import {RouteComponentProps} from 'react-router-dom'
 
 interface Props {
