+++ client/src/page/search/search.scss
